+++ packages/client/src/index.ts
@@ -1,6 +1,6 @@
-export * from "./balances";
+export * from "./getBalances";
 export * from "./client";
-export * from "./fees";
+export * from "./getFees";
 export * from "./pools";
 export * from "./prepare";
 export * from "./sendZETA";
